+++ lib/http/features/auto_inflate.rb
@@ -27,10 +27,11 @@ module HTTP
       #   feature.stream_for(connection)
       #
       # @param connection [HTTP::Connection]
+      # @param encoding [Encoding] encoding to use for the inflated body
       # @return [HTTP::Response::Body]
       # @api public
-      def stream_for(connection)
-        Response::Body.new(Response::Inflater.new(connection))
+      def stream_for(connection, encoding: Encoding::BINARY)
+        Response::Body.new(Response::Inflater.new(connection), encoding: encoding)
       end
 
       private
@@ -45,7 +46,7 @@ module HTTP
           headers:       response.headers,
           proxy_headers: response.proxy_headers,
           connection:    response.connection,
-          body:          stream_for(response.connection),
+          body:          stream_for(response.connection, encoding: response.body.encoding),
           request:       response.request
         }
       end